pyproject.toml
[project]
name = "apacheairflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow-providers-cncf-kubernetes>=10.8.0",
]
